+++ packages/admin/blocks-admin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @comet/blocks-admin
 
+## 4.4.3
+
+### Patch Changes
+
+-   @comet/admin@4.4.3
+-   @comet/admin-icons@4.4.3
+
 ## 4.4.2
 
 ### Patch Changes